organize open-digger python kernel code in a more object-oriented way
