+++ python/metrics/related_users.py
@@ -1,9 +1,11 @@
-from basic import getMergedConfig, getRepoWhereClauseForNeo4j, getTimeRangeWhereClauseForNeo4j
-import db.neo4j_driver as neo4j_driver
+from .basic import getMergedConfig, getRepoWhereClauseForNeo4j, getTimeRangeWhereClauseForNeo4j
+from db.neo4j_wrapper import Neo4jWrapper 
+neo4j = Neo4jWrapper()
 
-def getRelatedUsers(config):
-    config = getMergedConfig(config)
-    repoWhereClause = getRepoWhereClauseForNeo4j(config)
-    timeWhereClause = getTimeRangeWhereClauseForNeo4j(config, 'a')
-    query = 'MATCH (r:Repo)<-[a:ACTION]-(u:User) WHERE {} {} RETURN DISTINCT u.login AS user_login {};'.format(repoWhereClause + ' AND ' if repoWhereClause != None else '', timeWhereClause, 'LIMIT {}'.format(config.get('limit')) if config.get('limit') > 0 else '')
-    return neo4j_driver.query(query)
+class Relation():
+    def getRelatedUsers(config):
+        config = getMergedConfig(config)
+        repoWhereClause = getRepoWhereClauseForNeo4j(config)
+        timeWhereClause = getTimeRangeWhereClauseForNeo4j(config, 'a')
+        query = 'MATCH (r:Repo)<-[a:ACTION]-(u:User) WHERE {} {} RETURN DISTINCT u.login AS user_login {};'.format(repoWhereClause + ' AND ' if repoWhereClause != None else '', timeWhereClause, 'LIMIT {}'.format(config.get('limit')) if config.get('limit') > 0 else '')
+        return neo4j.query(query)
